Removed log preprocessor, added if around all
The client.Log function was compiled out using a preprocessor. The reasoning behind this was that expensive construction of strings were still happening. But it was a bit more annoying to enable logging in C# compared to our other libraries. I removed the need for the preprocessor and added if condition on each log instead, like we do in other libraries.