Fix for _localPath that causes a Format Exception Error
Error: FormatException: The input is not a valid Base-64 string as it contains a non-base 64 character, more than two padding characters, or an illegal character among the padding characters.

Reason: _localPath would direct the file that was created and then give the needed info for UploadFileFromMemory(). But once the response is received, the _localPath acts more of a GUID instead of a path. Not sure when this was changed.