Fix For Auth Advance As Anonymous Type (#235)
Previously due to InitializeIdentity defaulting to false, it would ALWAYS wipe the profileId that is stored. Also made a check to switch the externalId to anonymousId if the AuthenticateType == Anonymous

Unit test:
AuthenticateAdvance() as Anonymous type with forceCreate true, then log out and try to AuthenticateAdvance() again but with forceCreate false to ensure we're reading the same user. Once we get the result from the second authenticate, delete the user.

Adding a check for authentication token if anonymous

Making the reset stored id's into a pre success callback so if its successful